move TestsTable to its own folder
