+++ src/pages/task/TestsTable/TestsTable.test.tsx
@@ -3,8 +3,8 @@ import { render, cleanup } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
 import * as ReactDOM from "react-dom";
 import { MockedProvider } from "@apollo/react-testing";
-import { TestsTable } from "../../pages/task/TestsTable";
-import { GET_TASK_TESTS } from "../../gql/queries/get-task-tests";
+import { TestsTable } from ".";
+import { GET_TASK_TESTS } from "../../../gql/queries/get-task-tests";
 import { MemoryRouter, Route } from "react-router";
 import wait from "waait";
 import { fireEvent } from "@testing-library/react";
